`Refactored routes and components: removed JoinUs and Contributors, added OurTeam and Community`
